Complete bmp180GetTemperature, fixing data type of calib data

+++ drivers/Inc/bmp180.h
@@ -14,22 +14,22 @@
 typedef union{
 	uint16_t full[11];
 	struct{
-		uint16_t AC1;
-		uint16_t AC2;
-		uint16_t AC3;
+		int16_t AC1;
+		int16_t AC2;
+		int16_t AC3;
 		uint16_t AC4;
 		uint16_t AC5;
 		uint16_t AC6;
-		uint16_t B1;
-		uint16_t B2;
-		uint16_t MB;
-		uint16_t MC;
-		uint16_t MD;
+		int16_t B1;
+		int16_t B2;
+		int16_t MB;
+		int16_t MC;
+		int16_t MD;
 	}get;
 }BMP180_CALIBDATA;
 
 void bmp180GetCalibData(I2C_Handle_t* i2c, BMP180_CALIBDATA* calibData);
 uint32_t bmp180GetRawTemperature(I2C_Handle_t* i2c);
-float bmp180GetTemperature(I2C_Handle_t* i2c, BMP180_CALIBDATA* calibData);
+double  bmp180GetTemperature(I2C_Handle_t* i2c, BMP180_CALIBDATA* calibData);
 
 #endif /* DRIVERS_INC_BMP180_H_ */
